+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## NodeJS v2.0.0 - 2025-09-23
+- 🔥 **Breaking:** Correctly refering to bytes per sample as `bitDepth` now instead of `bitrate`.
+- 🔥 **Breaking:** `streamAudio` does only return the stream now and skip the voice check as this is done server side anyway.
+- The container enum now uses the correct order (as the server expects it) which means raw and wav have been swapped.
+- Added clearCache method to client.
+
+## Python v2.0.0 - 2025-09-23
+- 🔥 **Breaking:** Correctly refering to bytes per sample as `bit_depth` now instead of `bitrate`.
+- 🔥 **Breaking:** `stream_audio` does only return the stream now and skip the voice check as this is done server side anyway.
+- The container enum now uses the correct order (as the server expects it) which means raw and wav have been swapped.
+- Fixed the default ports to use 8423/8424 instead of 9423/9424.
+- Added clear_cache method to client.
+
 ## Rust v2.0.0 - 2025-09-22
 - 🔥 **Breaking:** Correctly refering to bytes per sample as `bit_depth` now instead of `bitrate`.
 - The container enum now uses the correct order (as the server expects it) which means raw and wav have been swapped.